Split MessageJumpToAbilitySection messages to top level
