+++ src/main/java/com/bobmowzie/mowziesmobs/server/ServerProxy.java
@@ -94,8 +94,8 @@ public class ServerProxy {
         this.registerMessage(MessageUseAbility.class, MessageUseAbility::serialize, MessageUseAbility::deserialize, new MessageUseAbility.Handler(), NetworkDirection.PLAY_TO_CLIENT);
         this.registerMessage(MessagePlayerUseAbility.class, MessagePlayerUseAbility::serialize, MessagePlayerUseAbility::deserialize, new MessagePlayerUseAbility.Handler(), NetworkDirection.PLAY_TO_SERVER);
         this.registerMessage(MessageInterruptAbility.class, MessageInterruptAbility::serialize, MessageInterruptAbility::deserialize, new MessageInterruptAbility.Handler(), NetworkDirection.PLAY_TO_CLIENT);
-        this.registerMessage(MessageJumpToAbilitySection.MessageJumpToAbilitySectionServerToClient.class, MessageJumpToAbilitySection.MessageJumpToAbilitySectionServerToClient::serialize, MessageJumpToAbilitySection.MessageJumpToAbilitySectionServerToClient::deserialize, new MessageJumpToAbilitySection.MessageJumpToAbilitySectionServerToClient.Handler(), NetworkDirection.PLAY_TO_CLIENT);
-        this.registerMessage(MessageJumpToAbilitySection.MessageJumpToAbilitySectionClientToServer.class, MessageJumpToAbilitySection.MessageJumpToAbilitySectionClientToServer::serialize, MessageJumpToAbilitySection.MessageJumpToAbilitySectionClientToServer::deserialize, new MessageJumpToAbilitySection.MessageJumpToAbilitySectionClientToServer.Handler(), NetworkDirection.PLAY_TO_SERVER);
+        this.registerMessage(MessageJumpToAbilitySectionServerToClient.class, MessageJumpToAbilitySectionServerToClient::serialize, MessageJumpToAbilitySectionServerToClient::deserialize, new MessageJumpToAbilitySectionServerToClient.Handler(), NetworkDirection.PLAY_TO_CLIENT);
+        this.registerMessage(MessageJumpToAbilitySectionClientToServer.class, MessageJumpToAbilitySectionClientToServer::serialize, MessageJumpToAbilitySectionClientToServer::deserialize, new MessageJumpToAbilitySectionClientToServer.Handler(), NetworkDirection.PLAY_TO_SERVER);
         this.registerMessage(MessageSculptorTrade.class, MessageSculptorTrade::serialize, MessageSculptorTrade::deserialize, new MessageSculptorTrade.Handler(), NetworkDirection.PLAY_TO_SERVER);
         this.registerMessage(MessageLinkEntities.class, MessageLinkEntities::serialize, MessageLinkEntities::deserialize, new MessageLinkEntities.Handler(), NetworkDirection.PLAY_TO_CLIENT);
         this.registerMessage(MessageUpdateBossBar.class, MessageUpdateBossBar::serialize, MessageUpdateBossBar::deserialize, new MessageUpdateBossBar.Handler(), NetworkDirection.PLAY_TO_CLIENT);
